+++ src/adapter/firefox/protocol.d.ts
@@ -145,9 +145,6 @@ declare namespace FirefoxDebugProtocol {
 		actor: string;
 		url: string;
 		type: number;
-		traits: {
-			doNotAttach?: boolean;
-		}
 	}
 
 	interface WorkerAttachedResponse extends TypedResponse {